# contracts/openapi/nexus.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentDock Nexus API",
    "version": "1.0.0",
    "description": "个人多设备 AgentDock 控制台的真实生产 HTTP 契约，覆盖设备、记忆、加密文件、备份状态和账号会话。"
  },
  "servers": [
    {
      "url": "/",
      "description": "当前 Nexus 实例。"
    }
  ],
  "paths": {
...
    "/v1/system/status": {
      "get": {
        "operationId": "getSystemStatus",
        "summary": "读取 Nexus 与 SQLite 状态",
        "responses": {
          "200": {
            "description": "成功。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatus"
                }
              }
            }
          },
          "400": {
            "description": "错误。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyErrorEnvelope"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "错误。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyErrorEnvelope"
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "错误。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyErrorEnvelope"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "错误。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyErrorEnvelope"
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "错误。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyErrorEnvelope"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/backup/status": {
      "get": {
        "operationId": "getBackupStatus",
        "summary": "读取 AgentDock 与 Nexus 备份状态",
        "responses": {
          "200": {
            "description": "成功。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupStatus"
                }
              }
            }
          },
          "400": {
            "description": "错误。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyErrorEnvelope"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "错误。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyErrorEnvelope"
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "错误。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyErrorEnvelope"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "错误。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyErrorEnvelope"
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "错误。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyErrorEnvelope"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/memories": {
      "get": {
        "operationId": "listMemories",
        "summary": "列出记忆条目",
        "responses": {
          "200": {
            "description": "成功。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "400": {
            "description": "错误。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyErrorEnvelope"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "错误。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyErrorEnvelope"
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "错误。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyErrorEnvelope"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "错误。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyErrorEnvelope"
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "错误。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyErrorEnvelope"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "writeMemory",
        "summary": "创建记忆条目",
        "responses": {
          "200": {
            "description": "成功。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "400": {
            "description": "错误。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyErrorEnvelope"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "错误。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyErrorEnvelope"
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "错误。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyErrorEnvelope"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "错误。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyErrorEnvelope"
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "错误。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyErrorEnvelope"
                    }
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryEntry"
              }
            }
          }
        }
      }
    },
    "/v1/memories/move": {
      "post": {
        "operationId": "moveMemory",
        "summary": "移动记忆条目",
        "responses": {
          "200": {
            "description": "成功。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "400": {
            "description": "错误。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyErrorEnvelope"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "错误。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyErrorEnvelope"
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "错误。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyErrorEnvelope"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "错误。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyErrorEnvelope"
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "错误。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyErrorEnvelope"
                    }
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonObject"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
      "JsonObject": {
        "type": "object",
        "description": "通用结构化对象。",
        "additionalProperties": true,
        "properties": {}
      },
      "ErrorDetail": {
        "type": "object",
        "description": "字段级错误详情。",
        "additionalProperties": false,
        "properties": {
          "field": {
            "type": "string",
            "description": "字段路径。"
          },
          "reason": {
            "type": "string",
            "description": "稳定原因标识。"
          },
          "message": {
            "type": "string",
            "description": "可读错误说明。"
          }
        },
        "required": [
          "reason",
          "message"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "description": "统一错误响应。",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string",
            "description": "稳定错误码。"
          },
          "message": {
            "type": "string",
            "description": "面向调用方的错误说明。"
          },
          "request_id": {
            "type": "string",
            "description": "请求关联 ID。"
          },
          "details": {
            "type": "array",
            "description": "可选字段级错误。",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ]
      },
      "LegacyErrorEnvelope": {
        "type": "object",
        "description": "Memory 与浏览器兼容接口使用的错误信封。",
        "additionalProperties": false,
        "properties": {
          "ok": {
            "type": "boolean",
            "description": "固定为 false。"
          },
          "error": {
            "type": "object",
            "description": "错误详情。",
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string",
                "description": "稳定错误码。"
              },
              "message": {
                "type": "string",
                "description": "可读错误说明。"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "ok",
          "error"
        ]
      },
...
      "SystemStatus": {
        "type": "object",
        "description": "Nexus 系统与数据存储状态。",
        "additionalProperties": false,
        "properties": {
          "ok": {
            "type": "boolean",
            "description": "系统是否健康。"
          },
          "service": {
            "type": "string",
            "description": "服务名称。"
          },
          "database": {
            "type": "string",
            "description": "SQLite 健康状态。"
          },
          "schema_version": {
            "type": "integer",
            "description": "数据库 Schema 版本。",
            "minimum": 0
          },
          "recall_root": {
            "type": "string",
            "description": "RecallDock 召回仓库路径。"
          },
          "memory_root": {
            "type": "string",
            "description": "旧字段：RecallDock 召回仓库路径。"
          },
          "artifact_root": {
            "type": "string",
            "description": "Artifact 密文存储路径。"
          }
        },
        "required": [
          "ok",
          "service",
          "database",
          "schema_version",
          "recall_root",
          "memory_root",
          "artifact_root"
        ]
      },
...
      "MemoryEntry": {
        "type": "object",
        "description": "Markdown 记忆条目。",
        "additionalProperties": false,
        "properties": {
          "path": {
            "type": "string",
            "description": "记忆相对路径。"
          },
          "content": {
            "type": "string",
            "description": "Markdown 或文本内容。"
          },
          "size_bytes": {
            "type": "integer",
            "description": "内容字节数。",
            "minimum": 0
          },
          "modified_at": {
            "type": "string",
            "description": "RFC 3339 UTC 时间。",
            "format": "date-time"
          }
        },
        "required": [
          "path"
        ]
      }
    }
  }
}